Fix simple polygon case.
We discounted a polygon if itsone of its vertices sat in the bounding
rectangle of one of the adjacent lines. This makes requires us to check
that the point is collinear too.

Create and test new onSegment template that checks that a point is both
collinear and in the bounding rectangle.